+++ db/migrate/20170322153936_add_order_to_reminders.rb
@@ -1,0 +1,5 @@
+class AddOrderToReminders < ActiveRecord::Migration[5.0]
+  def change
+    add_column :reminders, :order, :integer, default: 0
+  end
+end